[REFACTOR] 카테고리 선택 컴포넌트 파일 분리
