+++ src/constants/Food.js
@@ -3,21 +3,21 @@ export const category = {
   JAPANESE: '일식',
   CHINESE: '중식',
   WESTERN_ASIAN: '양식/아시안',
+  ALL: 'ALL',
   CHICKEN: '치킨',
   SNACKS: '분식',
   CAFE: '카페/디저트',
-  ALL: 'ALL',
   FASTFOOD: '패스트푸드',
 };
 
 export const dummyData = {
-  [category.KOREAN]: ["떡볶이", "김밥", "라면", "비빔밥", "육회비빔밥"],
-  [category.JAPANESE]: ["초밥", "라멘"],
-  [category.CHINESE]: ["짜장면", "짬뽕"],
-  [category.WESTERN_ASIAN]: ["파스타", "피자"],
-  [category.CHICKEN]: ["땅땅치킨", "교촌"],
-  [category.ALL]: ["햄버거", "핫도그"],
-  [category.SNACKS]: ["떡볶이", "핫도그","어묵"],
-  [category.CAFE]: ["아메리카노", "카페라떼", "카푸치노", "플랫화이트"],
-  [category.FASTFOOD]: ["삼각김밥", "도시락"],
+  [category.KOREAN]: ['떡볶이', '김밥', '라면', '비빔밥', '육회비빔밥'],
+  [category.JAPANESE]: ['초밥', '라멘'],
+  [category.CHINESE]: ['짜장면', '짬뽕'],
+  [category.WESTERN_ASIAN]: ['파스타', '피자'],
+  [category.CHICKEN]: ['땅땅치킨', '교촌'],
+  [category.ALL]: ['햄버거', '핫도그'],
+  [category.SNACKS]: ['떡볶이', '핫도그', '어묵'],
+  [category.CAFE]: ['아메리카노', '카페라떼', '카푸치노', '플랫화이트'],
+  [category.FASTFOOD]: ['삼각김밥', '도시락'],
 };
