+++ src/screens/Home.js
@@ -1,12 +1,11 @@
 import React from 'react';
-import { View, StyleSheet } from 'react-native';
 import Container from '../components/common/Container';
-import RandomButton from '../components/common/RandomButton';
 import DistanceButton from '../components/home/DistanceButton';
 import DistanceModal from '../components/home/DistanceModal';
+import CategorySelect from '../components/home/CategorySelect';
 import { getCurrentPosition } from '../components/common/GeolocationAPI';
 import { useState } from 'react';
-import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
+// import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
 
 const Home = () => {
   const [visible, setVisible] = useState(false);
@@ -31,40 +30,9 @@ const Home = () => {
         minute={minute}
         onSlidingComplete={(value) => setMinute(value)}
       />
-      <View style={styles.homegridContainer}>
-               <RandomButton category="분식"/>
-              <RandomButton category="일식" />
-              <RandomButton category="양식,아시안" />
-              <RandomButton category="치킨" />
-              <RandomButton category="ALL" />
-              <RandomButton category="한식" />
-              <RandomButton category="중식" />
-              <RandomButton category="카페" />
-              <RandomButton category="간편식" />
-            </View>
+      <CategorySelect />
     </Container>
   );
 };
-
-const styles = StyleSheet.create({
-  homegridContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    backgroundColor: '#FFFFFF',
-    borderRadius: 30,
-    alignItems: 'center',
-    justifyContent: 'center',
-    width:360,
-    height:360,
-    shadowColor: "#EAEAEA",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
-});
 
 export default Home;
